requirements.txt
pdfminer.six
pdf2image
Pillow
pytesseract
flask
requests
flask-cors